+++ CHANGES.md
@@ -1,6 +1,7 @@
 ## 4.0.3
 
 * `Cell::Partial` now does _append_ the global partial path to its `view_paths` instead of using `unshift` and thereby removing possible custom paths.
+* Adding `Cell::Translation` which allows using the `#t` helper. Thanks to @johnlane.
 
 ## 4.0.2